django-backend/requirements.txt
# Django
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.3.0,<4.4.0
python-dotenv>=1.0.0,<1.1.0
psycopg2-binary>=2.9.9,<3.0.0
gunicorn>=21.2.0,<22.0.0

# AI/ML
torch>=2.3.0,<2.4.0
torchvision>=0.18.0,<0.19.0
torchaudio>=2.3.0,<2.4.0
onnx>=1.16.0,<1.17.0
onnxruntime>=1.18.0,<1.19.0

# Media Processing
opencv-python>=4.9.0,<5.0.0
Pillow>=10.3.0,<11.0.0
ffmpeg-python>=0.2.0,<0.3.0
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0,<0.13.0

# API Integration
requests>=2.31.0,<2.32.0
aiohttp>=3.9.0,<3.10.0

# Utilities
numpy>=1.26.0,<2.0.0
scipy>=1.13.0,<1.14.0
tqdm>=4.66.0,<4.67.0
pydantic>=2.7.0,<2.8.0

# Redis for task queue
redis>=5.0.0,<5.1.0
celery>=5.3.0,<5.4.0

# Testing
pytest>=8.2.0,<8.3.0
pytest-django>=4.8.0,<4.9.0
